Update Local Master

+++ src/setup_payload/QRCodeSetupPayloadParser.h
@@ -46,8 +46,8 @@ public:
 
 private:
     CHIP_ERROR retrieveOptionalInfos(SetupPayload & outPayload, TLV::TLVReader & reader);
-    CHIP_ERROR populateTLV(SetupPayload & outPayload, const std::vector<uint8_t> & buf, int & index);
-    CHIP_ERROR parseTLVFields(chip::SetupPayload & outPayload, uint8_t * tlvDataStart, uint32_t tlvDataLengthInBytes);
+    CHIP_ERROR populateTLV(SetupPayload & outPayload, const std::vector<uint8_t> & buf, size_t & index);
+    CHIP_ERROR parseTLVFields(chip::SetupPayload & outPayload, uint8_t * tlvDataStart, size_t tlvDataLengthInBytes);
 };
 
 } // namespace chip
